Add Group subcomponents
Separator is not a subcomponent as it's defined outside of the group
and could be esasily forgotten when subcomponent approach is used.
